feat: DTO 생성 주체를 Repository 에서 Service 로 이전
