+++ src/databases/repositories/word.repository.ts
@@ -1,28 +1,14 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 
-import { plainToInstance } from 'class-transformer';
 import { DataSource, Repository } from 'typeorm';
 
-import { PaginationDto, PaginationMetaDto } from '#/common/dto/pagination.dto';
 import { RequestCreateWordDto } from '#/word/dto/create-word.dto';
 import { RequestUpdateWordDto } from '#/word/dto/update-word.dto';
-import {
-	RequestWordDetailDto,
-	ResponseWordDetailDto,
-} from '#/word/dto/word-detail.dto';
-import {
-	RequestWordListDto,
-	ResponseWordListDto,
-} from '#/word/dto/word-list.dto';
-import {
-	RequestWordSearchDto,
-	ResponseWordSearchDto,
-} from '#/word/dto/word-search.dto';
-import {
-	RequestWordUserLikeDto,
-	ResponseWordUserLikeDto,
-} from '#/word/dto/word-user-like.dto';
+import { RequestWordDetailDto } from '#/word/dto/word-detail.dto';
+import { RequestWordListDto } from '#/word/dto/word-list.dto';
+import { RequestWordSearchDto } from '#/word/dto/word-search.dto';
+import { RequestWordUserLikeDto } from '#/word/dto/word-user-like.dto';
 import { WORD_SORTING_TYPE } from '#/word/interface/word-list-sorting.interface';
 import { Word } from '#databases/entities/word.entity';
 
@@ -114,14 +100,7 @@ export class WordRepository {
 			queryBuilder.addSelect(['false AS isLike']);
 		}
 
-		const word = await queryBuilder.groupBy('word.id').getRawOne();
-
-		const responseWordDetailDto = plainToInstance(
-			ResponseWordDetailDto,
-			word,
-			{ excludeExtraneousValues: true },
-		);
-		return responseWordDetailDto;
+		return await queryBuilder.groupBy('word.id').getRawOne();
 	}
 
 	async findBySearchWord(requestWordSearchDto: RequestWordSearchDto) {
@@ -155,18 +134,7 @@ export class WordRepository {
 			queryBuilder.getCount(),
 		]);
 
-		const paginationMeta = new PaginationMetaDto({
-			paginationOption: requestWordSearchDto,
-			totalCount,
-		});
-
-		const responseWordSearchDto = plainToInstance(
-			ResponseWordSearchDto,
-			words,
-			{ excludeExtraneousValues: true },
-		);
-
-		return new PaginationDto(responseWordSearchDto, paginationMeta);
+		return { words, totalCount };
 	}
 
 	async findWithList(requestWordListDto: RequestWordListDto) {
@@ -209,18 +177,10 @@ export class WordRepository {
 			.createQueryBuilder('word')
 			.getCount();
 
-		const responseWordListDto = plainToInstance(
-			ResponseWordListDto,
+		return {
 			words,
-			{ excludeExtraneousValues: true },
-		);
-
-		const paginationMeta = new PaginationMetaDto({
-			paginationOption: requestWordListDto,
 			totalCount,
-		});
-
-		return new PaginationDto(responseWordListDto, paginationMeta);
+		};
 	}
 
 	async findUserLikeWord(requestWordListDto: RequestWordUserLikeDto) {
@@ -245,16 +205,9 @@ export class WordRepository {
 			.take(requestWordListDto.limit)
 			.getManyAndCount();
 
-		const paginationMeta = new PaginationMetaDto({
-			paginationOption: requestWordListDto,
-			totalCount,
-		});
-
-		const responseWordListDto = plainToInstance(
-			ResponseWordUserLikeDto,
+		return {
 			words,
-		);
-
-		return new PaginationDto(responseWordListDto, paginationMeta);
+			totalCount,
+		};
 	}
 }
