+++ src/word/word.service.ts
@@ -3,15 +3,25 @@ import { Cron, CronExpression } from '@nestjs/schedule';
 
 import { plainToInstance } from 'class-transformer';
 
+import { PaginationDto, PaginationMetaDto } from '#/common/dto/pagination.dto';
 import { SpreadSheetService } from '#/spread-sheet/spread-sheet.service';
 import { WordRepository } from '#databases/repositories/word.repository';
 
 import { RequestCreateWordDto } from './dto/create-word.dto';
 import { RequestUpdateWordDto } from './dto/update-word.dto';
-import { RequestWordDetailDto } from './dto/word-detail.dto';
-import { RequestWordListDto } from './dto/word-list.dto';
-import { RequestWordSearchDto } from './dto/word-search.dto';
-import { RequestWordUserLikeDto } from './dto/word-user-like.dto';
+import {
+	RequestWordDetailDto,
+	ResponseWordDetailDto,
+} from './dto/word-detail.dto';
+import { RequestWordListDto, ResponseWordListDto } from './dto/word-list.dto';
+import {
+	RequestWordSearchDto,
+	ResponseWordSearchDto,
+} from './dto/word-search.dto';
+import {
+	RequestWordUserLikeDto,
+	ResponseWordUserLikeDto,
+} from './dto/word-user-like.dto';
 
 @Injectable()
 export class WordService {
@@ -102,12 +112,33 @@ export class WordService {
 		return await this.updateWordList();
 	}
 
-	async getWordList(wordListDto: RequestWordListDto) {
-		return await this.wordRepository.findWithList(wordListDto);
+	async getWordList(requestWordListDto: RequestWordListDto) {
+		const { words, totalCount } =
+			await this.wordRepository.findWithList(requestWordListDto);
+
+		const responseWordListDto = plainToInstance(ResponseWordListDto, words);
+		const paginationMeta = new PaginationMetaDto({
+			paginationOption: requestWordListDto,
+			totalCount,
+		});
+
+		return new PaginationDto(responseWordListDto, paginationMeta);
 	}
 
-	async getWordUserLike(wordUserLikeDto: RequestWordUserLikeDto) {
-		return await this.wordRepository.findUserLikeWord(wordUserLikeDto);
+	async getWordUserLike(requestWordUserLikeDto: RequestWordUserLikeDto) {
+		const { words, totalCount } =
+			await this.wordRepository.findUserLikeWord(requestWordUserLikeDto);
+
+		const responseWordUserLikeListDto = plainToInstance(
+			ResponseWordUserLikeDto,
+			words,
+		);
+		const paginationMeta = new PaginationMetaDto({
+			paginationOption: requestWordUserLikeDto,
+			totalCount,
+		});
+
+		return new PaginationDto(responseWordUserLikeListDto, paginationMeta);
 	}
 
 	async getWordById(wordDetailDto: RequestWordDetailDto) {
@@ -118,10 +149,31 @@ export class WordService {
 			throw new BadRequestException(
 				`해당 ID 를 가진 Word 가 존재하지 않습니다.`,
 			);
-		return word;
+
+		const responseWordDetailDto = plainToInstance(
+			ResponseWordDetailDto,
+			word,
+			{ excludeExtraneousValues: true },
+		);
+
+		return responseWordDetailDto;
 	}
 
 	async getWordByKeyword(requestWordSearchDto: RequestWordSearchDto) {
-		return await this.wordRepository.findBySearchWord(requestWordSearchDto);
+		const { words, totalCount } =
+			await this.wordRepository.findBySearchWord(requestWordSearchDto);
+
+		const paginationMeta = new PaginationMetaDto({
+			paginationOption: requestWordSearchDto,
+			totalCount,
+		});
+
+		const responseWordSearchDto = plainToInstance(
+			ResponseWordSearchDto,
+			words,
+			{ excludeExtraneousValues: true },
+		);
+
+		return new PaginationDto(responseWordSearchDto, paginationMeta);
 	}
 }
